gen_data/requirements.txt
albumentations==1.4.0
opencv-python-headless>=4.9.0
Pillow>=10.0
PyYAML>=6.0
rich>=13.7
loguru>=0.7
tqdm>=4.66
pytest>=8.1
# --- диффузионные модели ---
diffusers==0.28.0
transformers>=4.40
accelerate>=0.28
# для Yandex API
requests>=2.32